# config/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1",
    "title": "Free Busy Service",
    "description": "This API will fetch details about calendar id free or busy status.",
    "contact": {}
  },
  "paths": {
    "/getFreeBusy": {
      "get": {
        "description": "Provide start date and calender id. make sure that make sure Calendar\nowner shared calendar access to you",
        "summary": "Get user status of the availability ",
        "tags": [
          "Free Busy Service"
        ],
        "operationId": "get_free_busy",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "description": "Start date of the calendar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "calendarID",
            "in": "query",
            "required": true,
            "description": "enter Calendar Id. it is Gmail Id of the person",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success response"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "freebusy_auth": [
              "write:calendar",
              "read:calendar"
            ]
          }
        ]
      }
    }
  },
  "tags": [
    {
      "name": "Free Busy Service",
      "description": "This is foundation API and will be used to fetch availability of person.\nJust fill the request body right."
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "freebusy_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://accounts.google.com/o/oauth2/auth",
            "scopes": {
              "https://www.googleapis.com/auth/calendar": "modify calendar",
              "https://www.googleapis.com/auth/calendar.readonly": "read your calendar"
            }
          }
        }
      }
    }
  }
}